Refactor passenger count input in OrderRide page.
